Fix 14 patient flow issues from audit
- Delete orphaned /book page and dead POST /api/matches route
- Prevent rescheduling appointments to past dates
- Simplify pending booking count filter (use user_id OR email, not both)
- Move OTP hash clearing to after session token generation (retry-safe)
- Clear intake form draft after lead creation to prevent duplicates
- Show error when messaging without an active lead
- Fix "My Account" nav to route clinic users to /clinic
- Add session expiry detection banner on patient dashboard
- Limit re-requests to 3 per clinic after declines
- Replace fixed 1500ms post-OTP delay with session polling
- Remove auto-complete logic (clinic must confirm attendance)
- Hide quick prompts after 2+ messages on desktop (match mobile)
- Standardize date formatting to short format across views

+++ components/match/booking-card.tsx
@@ -365,7 +365,7 @@ export function BookingCard({
               availableHours={clinic.available_hours || ["09:00", "10:00", "11:00", "12:00", "13:00", "14:00", "15:00", "16:00", "17:00"]}
               acceptsSameDay={clinic.accepts_same_day || false}
               onSelectSlot={(date, time) => {
-                const dateLabel = date.toLocaleDateString("en-GB", { weekday: "long", day: "numeric", month: "long" })
+                const dateLabel = date.toLocaleDateString("en-GB", { weekday: "short", day: "numeric", month: "short" })
                 const msg = `Hi! I'd like to request an appointment on ${dateLabel} at ${time}. Would this time be available?`
                 setPendingAppointment({ message: msg, dateLabel, timeLabel: time })
               }}
@@ -432,7 +432,7 @@ export function BookingCard({
           {/* Appointment status banner */}
           {(appointmentRequested || bookingStatus === "confirmed" || bookingStatus === "declined" || bookingStatus === "cancelled" || bookingStatus === "completed") && (() => {
             const formattedBookingDate = bookingDate
-              ? new Date(bookingDate).toLocaleDateString("en-GB", { weekday: "long", day: "numeric", month: "long" })
+              ? new Date(bookingDate).toLocaleDateString("en-GB", { weekday: "short", day: "numeric", month: "short" })
               : null
             const formattedBookingTime = bookingTime
               ? (HOURLY_SLOTS.find((s) => s.key === bookingTime)?.label || bookingTime)
